fastcrypto_vdf::class_group
Module discriminant

Discriminants of quadratic forms are negative primes which is 1 mod 8. This module contains a type to represent discriminants and methods to create them.

Structs

DISCRIMINANT_3072

Fixed 3072 bit discriminant. Generated from the seed [1,2,3] using Discriminant::from_seed.

Discriminant

A discriminant for an imaginary class group. The discriminant is a negative integer congruent to 1 mod 8.