Struct fastcrypto_vdf::class_group::QuadraticForm

pub struct QuadraticForm {
    pub a: BigInt,
    pub b: BigInt,
    pub c: BigInt,
    /* private fields */
}

A binary quadratic form, (a, b, c) for arbitrary integers a, b, and c.

The partial_gcd_limit variable must be equal to |discriminant|^{1/4} and is used to speed up the composition algorithm.

Fields
a: BigIntb: BigIntc: BigInt
Implementations

impl QuadraticForm

pub fn hash_to_group( seed: &[u8], discriminant: &Discriminant, k: u16 ) -> FastCryptoResult<Self>

Generate a random quadratic form from a seed with the given discriminant. This method is deterministic and it is a random oracle on a large subset of the class group, namely the group elements whose a coordinate is a product of k primes all smaller than (sqrt(|discriminant|)/2)^{1/k}.

Increasing k speeds-up the function (at least up to some break even point), but it also decreases the size of the range of the hash function, so k must be picked no larger than the k computed in [largest_allowed_k]. If it is larger, an [InvalidInput] error is returned. If in doubt, use the [hash_to_group_with_default_parameters] instead.

pub fn hash_to_group_with_default_parameters( seed: &[u8], discriminant: &Discriminant ) -> FastCryptoResult<Self>

Generate a random quadratic form from a seed with the given discriminant. This method is deterministic, and it is a random oracle on a large subset of the class group. This method picks a default k parameter and calls the hash_to_group function with this k.

impl QuadraticForm

pub fn is_normal(&self) -> bool

Return true if this form is in normal form: -a < b <= a.

pub fn normalize(&mut self)

Return a normalized form equivalent to this quadratic form. See QuadraticForm::is_normal.

pub fn is_reduced(&self) -> bool

Return true if this form is reduced: A form is reduced if it is normal (see QuadraticForm::is_normal) and a <= c and if a == c then b >= 0.

pub fn reduce(&mut self)

Return a reduced form (see QuadraticForm::is_reduced) equivalent to this quadratic form.

impl QuadraticForm

pub fn from_a_b_discriminant( a: BigInt, b: BigInt, discriminant: &Discriminant ) -> FastCryptoResult<Self>

Create a new quadratic form given only the a and b coefficients and the discriminant.

pub fn from_bytes( bytes: &[u8], discriminant: &Discriminant ) -> FastCryptoResult<Self>

Create a new quadratic form from a serialization. The format is a byte array with the following structure:

a_len (2 bytes, big endian) | a as unsigned big endian bytes | b_len (2 bytes, big endian) | b as signed big endian bytes

The c coefficient is computed from a and b and the discriminant. If the format is not followed or if the coefficients do not form a valid quadratic form, an error is returned.

See also QuadraticForm::to_bytes.

pub fn generator(discriminant: &Discriminant) -> Self

Return a generator (or, more precisely, an element with a presumed large order) in a class group with a given discriminant which is 1 mod 8. We use the element (2, 1, c) where c is determined from the discriminant.

pub fn discriminant(&self) -> Discriminant

Compute the discriminant b^2 - 4ac for this quadratic form.

pub fn compose(&self, rhs: &QuadraticForm) -> QuadraticForm

Compute the composition of this quadratic form with another quadratic form.