Function fastcrypto_vdf::hash_prime::verify_prime

pub fn verify_prime<P: PrimalityCheck>(
    seed: &[u8],
    length_in_bytes: usize,
    bitmask: &[usize],
    index: usize
) -> FastCryptoResult<()>

Verify that the given prime is a prime and has the given index in the hash prime iterator.