Module fastcrypto_vdf::math::extended_gcd

This module contains an implementation of the extended Euclidean algorithm for BigInt`s. Besides the gcd and the Bezout coefficients, it also returns the quotients of the two inputs divided by the GCD since these are often used, for example in the NUCOMP and NUDPL algorithms, and come out for free while computing the Bezout coefficients.

Structs

EuclideanAlgorithmOutput

The output of the extended Euclidean algorithm on inputs a and b: The Bezout coefficients x and y such that ax + by = gcd. The quotients a / gcd and b / gcd are also returned.

Functions

extended_euclidean_algorithm

Compute the greatest common divisor gcd of a and b. The output also returns the Bezout coefficients x and y such that ax + by = gcd and also the quotients a / gcd and b / gcd.