Trait fastcrypto_vdf::math::parameterized_group::ParameterizedGroupElement

pub trait ParameterizedGroupElement: Sized + Clone + for<'a> Add<&'a Self, Output = Self> + Add<Output = Self> + Neg + Eq + Doubling {
    type ParameterType: Parameter;
    type ScalarType: From<u64>;

    // Required methods
    fn zero(parameter: &Self::ParameterType) -> Self;
    fn is_in_group(&self, parameter: &Self::ParameterType) -> bool;
}

Trait implemented by elements of an additive group where the group is parameterized, for example by the modulus in case of the group being Z mod N or the discriminant in case of class groups.

Required Associated Types

type ParameterType: Parameter

The type of the parameter which uniquely defines this group.

type ScalarType: From<u64>

Integer type used for multiplication.

Required Methods

fn zero(parameter: &Self::ParameterType) -> Self

Return an instance of the identity element in this group.

fn is_in_group(&self, parameter: &Self::ParameterType) -> bool

Returns true if this is an element of the group defined by parameter.

Object Safety

This trait is not object safe.