Module fastcrypto_vdf::vdf::wesolowski

Structs

FastVerifier

A faster method of verification which uses fast multi-scalar multiplication. The scalar size for the scalar multiplier M must be larger enough to hold the challenge in the Fiat-Shamir construction F.

StrongFiatShamir

Implementation of the Fiat-Shamir challenge generation for usage with Wesolowski’s VDF construction. The implementation is strong, meaning that all public parameters are used in the challenge generation. See https://eprint.iacr.org/2023/691.

WeakFiatShamir

Implementation of the Fiat-Shamir challenge generation compatible with chiavdf. Note that this implementation is weak, meaning that not all public parameters are used in the challenge generation. This is not secure if an adversary can influence the public parameters. See https://eprint.iacr.org/2023/691.

WesolowskisVDF

An implementation of Wesolowski’s VDF construction (https://eprint.iacr.org/2018/623) over a group of unknown order.

Constants

CHALLENGE_SIZE

Default size in bytes of the Fiat-Shamir challenge used in proving and verification (same as chiavdf).

Traits

FiatShamir

Type Aliases

StrongVDF

Implementation of Wesolowski’s VDF construction over a group of unknown order using a strong Fiat-Shamir implementation.

StrongVDFVerifier

WeakVDF

Implementation of Wesolowski’s VDF construction over a group of unknown order using the Fiat-Shamir construction from chiavdf (https://github.com/Chia-Network/chiavdf).

WeakVDFVerifier