Trait fastcrypto_vdf::math::parameterized_group::ParameterizedGroupElement

pub trait ParameterizedGroupElement: Sized + Clone + for<'a> Add<&'a Self, Output = Self> + Add<Output = Self> + for<'a> Mul<&'a Self::ScalarType, Output = Self> + Neg + Eq + Doubling {
    type ParameterType: Parameter;
    type ScalarType: From<u64>;

    // Required methods
    fn zero(parameters: &Self::ParameterType) -> Self;
    fn mul(&self, scale: &Self::ScalarType) -> Self;
    fn parameter(&self) -> Self::ParameterType;

    // Provided method
    fn same_group(&self, other: &Self) -> bool { ... }
}

Trait implemented by elements of an additive group where the group is parameterized, for example by the modulus in case of the group being Z mod N or the discriminant in case of class groups.

Required Associated Types

type ParameterType: Parameter

The type of the parameter which uniquely defines this group.

type ScalarType: From<u64>

Integer type used for multiplication.

Required Methods

fn zero(parameters: &Self::ParameterType) -> Self

Return an instance of the identity element in this group.

fn mul(&self, scale: &Self::ScalarType) -> Self

Compute scale * self.

fn parameter(&self) -> Self::ParameterType

Return the parameter for the group this element belongs to.

Provided Methods

fn same_group(&self, other: &Self) -> bool

Check whether this group element is in the same group as other.

Object Safety

This trait is not object safe.