fastcrypto_vdf::math::parameterized_group

Trait ParameterizedGroupElement

pub trait ParameterizedGroupElement:
    Sized
    + Clone
    + for<'a> Add<&'a Self, Output = Self>
    + Eq
    + Doubling
    + Serialize {
    type ParameterType: Serialize;

    // Required methods
    fn zero(parameter: &Self::ParameterType) -> Self;
    fn is_in_group(&self, parameter: &Self::ParameterType) -> bool;

    // Provided method
    fn multiply(
        &self,
        scalar: &BigUint,
        parameter: &Self::ParameterType,
    ) -> Self { ... }
}

Trait implemented by elements of an additive group where the group is parameterized, for example by the modulus in case of the group being Z mod N or the discriminant in case of class groups.

Required Associated Types

type ParameterType: Serialize

The type of the parameter which uniquely defines this group.

Required Methods

fn zero(parameter: &Self::ParameterType) -> Self

Return an instance of the identity element in this group.

fn is_in_group(&self, parameter: &Self::ParameterType) -> bool

Returns true if this is an element of the group defined by parameter.

Provided Methods

fn multiply(&self, scalar: &BigUint, parameter: &Self::ParameterType) -> Self

Compute self * scalar.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.