fastcrypto_vdf::vdf

Trait VDF

pub trait VDF {
    type InputType;
    type OutputType;
    type ProofType;

    // Required methods
    fn evaluate(
        &self,
        input: &Self::InputType,
    ) -> FastCryptoResult<(Self::OutputType, Self::ProofType)>;
    fn verify(
        &self,
        input: &Self::InputType,
        output: &Self::OutputType,
        proof: &Self::ProofType,
    ) -> FastCryptoResult<()>;
}

This represents a Verifiable Delay Function (VDF) construction.

Required Associated Types

type InputType

The type of the input to the VDF.

type OutputType

The type of the output from the VDF.

type ProofType

The type of the proof of correctness for this VDF.

Required Methods

fn evaluate( &self, input: &Self::InputType, ) -> FastCryptoResult<(Self::OutputType, Self::ProofType)>

Evaluate this VDF and return the output and a proof of correctness.

fn verify( &self, input: &Self::InputType, output: &Self::OutputType, proof: &Self::ProofType, ) -> FastCryptoResult<()>

Verify the output and proof from a VDF.